Extract common replication stream functionality
