+++ src/common/src/main/scala/osmesa/common/streaming/ChangesetsMicroBatchReader.scala
@@ -2,18 +2,15 @@ package osmesa.common.streaming
 
 import java.net.URI
 import java.util
-import java.util.Optional
 
-import org.apache.spark.internal.Logging
 import org.apache.spark.sql.Row
 import org.apache.spark.sql.sources.v2.DataSourceOptions
-import org.apache.spark.sql.sources.v2.reader.streaming.{MicroBatchReader, Offset}
+import org.apache.spark.sql.sources.v2.reader.streaming.Offset
 import org.apache.spark.sql.sources.v2.reader.{DataReader, DataReaderFactory}
 import org.apache.spark.sql.types._
 import osmesa.common.model.Changeset
 
 import scala.collection.JavaConverters._
-import scala.compat.java8.OptionConverters._
 
 case class ChangesetsStreamBatchTask(baseURI: URI, start: SequenceOffset, end: SequenceOffset)
     extends DataReaderFactory[Row] {
@@ -22,35 +19,12 @@ case class ChangesetsStreamBatchTask(baseURI: URI, start: SequenceOffset, end: S
 }
 
 class ChangesetsStreamBatchReader(baseURI: URI, start: SequenceOffset, end: SequenceOffset)
-    extends DataReader[Row]
-    with Logging {
-  private var currentOffset = start
-  private var index = -1
-  private var changesets: Vector[Changeset] = _
-
-  override def next(): Boolean = {
-    index += 1
-
-    if (Option(changesets).isEmpty) {
-      // initialize changesets from the starting sequence
-      changesets = ChangesetsSource.getSequence(baseURI, currentOffset.sequence).toVector
-    }
-
-    // fetch next batch of changesets if necessary
-    // this is a loop in case sequences contain no changesets
-    while (index >= changesets.length && currentOffset + 1 < end) {
-      // fetch next sequence
-      currentOffset += 1
-      changesets = ChangesetsSource.getSequence(baseURI, currentOffset.sequence).toVector
-
-      index = 0
-    }
-
-    currentOffset < end && index < changesets.length
-  }
+    extends ReplicationStreamBatchReader[Changeset](baseURI, start, end) {
+  override def getSequence(baseURI: URI, sequence: Int): Seq[Changeset] =
+    ChangesetsSource.getSequence(baseURI, sequence)
 
   override def get(): Row = {
-    val changeset = changesets(index)
+    val changeset = items(index)
 
     Row(
       currentOffset.sequence,
@@ -74,10 +48,7 @@ class ChangesetsStreamBatchReader(baseURI: URI, start: SequenceOffset, end: Sequ
 }
 
 class ChangesetsMicroBatchReader(options: DataSourceOptions, checkpointLocation: String)
-    extends MicroBatchReader
-    with Logging {
-  val DefaultBatchSize: Int = 100
-
+    extends ReplicationStreamMicroBatchReader(options, checkpointLocation) {
   // TODO extract me
   val ChangesetSchema = StructType(
     StructField("sequence", IntegerType) ::
@@ -103,70 +74,11 @@ class ChangesetsMicroBatchReader(options: DataSourceOptions, checkpointLocation:
       .get("base_uri")
       .orElse("https://planet.osm.org/replication/changesets/"))
 
-  private val batchSize = options
-    .get("batch_size")
-    .asScala
-    .map(s => s.toInt)
-    .getOrElse(DefaultBatchSize)
-
-  private var start: Option[SequenceOffset] = options
-    .get("start_sequence")
-    .asScala
-    .map(s => SequenceOffset(s.toInt))
-
-  private var end: Option[SequenceOffset] = options
-    .get("end_sequence")
-    .asScala
-    .map(s => SequenceOffset(s.toInt))
-
-  override def setOffsetRange(start: Optional[Offset], end: Optional[Offset]): Unit = {
-    // TODO memoize this, valid for 30s at a time
-    val currentOffset = ChangesSource.createOffsetForCurrentSequence(baseURI)
-
-    this.start = Some(
-      start.asScala
-        .map(_.asInstanceOf[SequenceOffset])
-        .getOrElse {
-          this.start.getOrElse {
-            currentOffset - 1
-          }
-        })
-
-    this.end = Some(
-      end.asScala
-        .map(_.asInstanceOf[SequenceOffset])
-        .getOrElse {
-          val next = this.end.map(_ + 1).getOrElse {
-            this.start.get + 1
-          }
-
-          if (currentOffset > next) {
-            SequenceOffset(math.min(currentOffset.sequence, next.sequence + batchSize))
-          } else {
-            next
-          }
-        })
-  }
-
-  override def getStartOffset: Offset = {
-    start.getOrElse {
-      throw new IllegalStateException("start offset not set")
-    }
-  }
-
-  override def getEndOffset: Offset = {
-    end.getOrElse {
-      throw new IllegalStateException("end offset not set")
-    }
-  }
-
-  override def deserializeOffset(json: String): Offset =
-    SequenceOffset(json.toInt)
+  override def getCurrentOffset: SequenceOffset =
+    ChangesetsSource.createOffsetForCurrentSequence(baseURI)
 
   override def commit(end: Offset): Unit =
     logInfo(s"Changeset sequence ${end} processed.")
-
-  override def stop(): Unit = Unit
 
   override def readSchema(): StructType = ChangesetSchema
