+++ src/common/src/main/scala/osmesa/common/streaming/ChangesetsSource.scala
@@ -45,7 +45,7 @@ object ChangesetsSource extends Logging {
   private[streaming] def createOffsetForCurrentSequence(baseURI: URI): SequenceOffset =
     SequenceOffset(getCurrentSequence(baseURI))
 
-  def getSequence(baseURI: URI, sequence: Long): Seq[Changeset] = {
+  def getSequence(baseURI: URI, sequence: Int): Seq[Changeset] = {
     val s = f"$sequence%09d".toArray
     val path =
       s"${s.slice(0, 3).mkString}/${s.slice(3, 6).mkString}/${s.slice(6, 9).mkString}.osm.gz"
